+++ kernel/math/math.h
@@ -3,8 +3,12 @@
 
 #include <stdint.h>
 
-uint64_t divideRoundUp1(const uint64_t a, const uint64_t b);
-uint64_t min(const uint64_t a, const uint64_t b);
-uint64_t max(const uint64_t a, const uint64_t b);
+int64_t divide_round_up(const int64_t a, const int64_t b);
+int64_t min(const int64_t a, const int64_t b);
+int64_t max(const int64_t a, const int64_t b);
+
+uint64_t abs(const int64_t n);
+int64_t clamp(const int64_t n, const int64_t min, const int64_t max);
+bool is_power_of_two(const int64_t n);
 
 #endif