+++ kernel/types/char/char.h
@@ -1,0 +1,13 @@
+#ifndef CHAR_H
+#define CHAR_H
+
+bool is_digit(const char c);
+bool is_alpha(const char c);
+bool is_upper_alpha(const char c);
+bool is_lower_alpha(const char c);
+bool is_whitespace(const char c);
+
+char to_upper(const char c);
+char to_lower(const char c);
+
+#endif